+++ iris/service/content/file/__init__.py
@@ -1,0 +1,5 @@
+from .document import File  # noqa
+
+
+def includeme(config):
+    pass