+++ iris/service/content/file/document.py
@@ -1,0 +1,20 @@
+from lovely.esdb.document import Document
+from lovely.esdb.properties import Property
+
+from iris.service.db.dc import dc_defaults, DC_CREATED, DC_MODIFIED
+
+
+class File(Document):
+    """A file in the database
+
+    TODO: needs to be completed
+    """
+
+    INDEX = 'files'
+
+    id = Property(primary_key=True)
+
+    dc = Property(
+        default=dc_defaults(DC_CREATED, DC_MODIFIED),
+        doc="Dublin Core data."
+    )