+++ iris/service/content/weblocation/__init__.py
@@ -1,0 +1,5 @@
+from .document import WebLocation  # noqa
+
+
+def includeme(config):
+    pass